requirements.txt
opencv-python
tensorflow
numpy
pandas
scikit-learn
requests
pyarrow
pillow
torch
torchvision
